requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-generativeai==0.8.3
scikit-learn==1.5.2
numpy==1.26.4
python-dotenv==1.0.1
groq==0.11.0